Move Coordinate class to separate module
