+++ pants-plugins/internal_plugins/test_lockfile_fixtures/lockfile_fixture.py
@@ -9,10 +9,12 @@ from typing import Iterable
 
 from _pytest.fixtures import FixtureRequest
 
-from pants.jvm.resolve.common import ArtifactRequirement, ArtifactRequirements, Coordinate
 from pants.jvm.resolve.coursier_fetch import CoursierResolvedLockfile
 from pants.jvm.resolve.lockfile_metadata import LockfileContext
 from pants.util.docutil import bin_name
+from pants.jvm.resolve.common import ArtifactRequirement
+from pants.jvm.resolve.common import ArtifactRequirements
+from pants.jvm.resolve.coordinate import Coordinate
 
 
 @dataclass(frozen=True)
