+++ src/python/pants/backend/openapi/codegen/java/rules.py
@@ -58,11 +58,11 @@ from pants.jvm.dependency_inference.artifact_mapper import (
     AllJvmArtifactTargets,
     find_jvm_artifacts_or_raise,
 )
-from pants.jvm.resolve.common import Coordinate
 from pants.jvm.subsystems import JvmSubsystem
 from pants.jvm.target_types import JvmResolveField
 from pants.source.source_root import SourceRoot, SourceRootRequest
 from pants.util.logging import LogLevel
+from pants.jvm.resolve.coordinate import Coordinate
 
 
 class GenerateJavaFromOpenAPIRequest(GenerateSourcesRequest):
