+++ src/python/pants/backend/scala/util_rules/versions.py
@@ -7,8 +7,8 @@ from dataclasses import dataclass
 from enum import Enum
 
 from pants.engine.rules import collect_rules, rule
-from pants.jvm.resolve.common import Coordinate
 from pants.util.strutil import softwrap
+from pants.jvm.resolve.coordinate import Coordinate
 
 
 class InvalidScalaVersion(ValueError):
