+++ src/python/pants/jvm/resolve/coursier_fetch_filter_test.py
@@ -9,9 +9,10 @@ from unittest import mock
 import pytest
 
 from pants.engine.fs import EMPTY_DIGEST
-from pants.jvm.resolve.common import Coordinate, Coordinates
 from pants.jvm.resolve.coursier_fetch import CoursierLockfileEntry, CoursierResolvedLockfile
 from pants.jvm.resolve.key import CoursierResolveKey
+from pants.jvm.resolve.coordinate import Coordinate
+from pants.jvm.resolve.coordinate import Coordinates
 
 coord1 = Coordinate("test", "art1", "1.0.0")
 coord2 = Coordinate("test", "art2", "1.0.0")
