+++ src/python/pants/jvm/resolve/coursier_fetch_integration_test.py
@@ -8,26 +8,31 @@ import textwrap
 
 import pytest
 
-from pants.base.specs import RawSpecs, RecursiveGlobSpec
-from pants.core.util_rules import config_files, source_files
 from pants.engine.fs import FileDigest
 from pants.engine.internals.scheduler import ExecutionError
 from pants.engine.process import ProcessExecutionFailure
 from pants.engine.target import Targets
 from pants.jvm.compile import ClasspathEntry
-from pants.jvm.resolve.common import (
-    ArtifactRequirement,
-    ArtifactRequirements,
-    Coordinate,
-    Coordinates,
-)
-from pants.jvm.resolve.coursier_fetch import CoursierLockfileEntry, CoursierResolvedLockfile
 from pants.jvm.resolve.coursier_fetch import rules as coursier_fetch_rules
-from pants.jvm.target_types import JvmArtifactJarSourceField, JvmArtifactTarget
 from pants.jvm.testutil import maybe_skip_jdk_test
 from pants.jvm.util_rules import ExtractFileDigest
 from pants.jvm.util_rules import rules as util_rules
-from pants.testutil.rule_runner import PYTHON_BOOTSTRAP_ENV, QueryRule, RuleRunner, engine_error
+from pants.jvm.resolve.common import ArtifactRequirement
+from pants.jvm.resolve.common import ArtifactRequirements
+from pants.jvm.resolve.coordinate import Coordinate
+from pants.jvm.resolve.coordinate import Coordinates
+from pants.base.specs import RawSpecs
+from pants.base.specs import RecursiveGlobSpec
+from pants.core.util_rules import config_files
+from pants.core.util_rules import source_files
+from pants.jvm.resolve.coursier_fetch import CoursierLockfileEntry
+from pants.jvm.resolve.coursier_fetch import CoursierResolvedLockfile
+from pants.jvm.target_types import JvmArtifactJarSourceField
+from pants.jvm.target_types import JvmArtifactTarget
+from pants.testutil.rule_runner import PYTHON_BOOTSTRAP_ENV
+from pants.testutil.rule_runner import QueryRule
+from pants.testutil.rule_runner import RuleRunner
+from pants.testutil.rule_runner import engine_error
 
 HAMCREST_COORD = Coordinate(
     group="org.hamcrest",
@@ -253,7 +258,7 @@ def test_resolve_with_packaging(rule_runner: RuleRunner) -> None:
 @maybe_skip_jdk_test
 def test_resolve_with_broken_url(rule_runner: RuleRunner) -> None:
     coordinate = ArtifactRequirement(
-        coordinate=Coordinate(
+        Coordinate(
             group="org.hamcrest",
             artifact="hamcrest-core",
             version="1.3_inexplicably_wrong",  # if the group/artifact/version is real, coursier will fallback
@@ -273,7 +278,7 @@ def test_resolve_with_broken_url(rule_runner: RuleRunner) -> None:
 @maybe_skip_jdk_test
 def test_resolve_with_working_url(rule_runner: RuleRunner) -> None:
     requirement = ArtifactRequirement(
-        coordinate=Coordinate(
+        Coordinate(
             group="apache-commons-local",
             artifact="commons-collections",
             version="1.0.0_JAR_LOCAL",
@@ -335,7 +340,7 @@ def test_resolve_with_a_jar(rule_runner: RuleRunner) -> None:
     jar_field = jeremy_target[JvmArtifactJarSourceField]
 
     requirement = ArtifactRequirement(
-        coordinate=Coordinate(
+        Coordinate(
             group="jeremy",
             artifact="jeremy",
             version="4.13.2",
@@ -654,12 +659,7 @@ def test_transitive_excludes(rule_runner: RuleRunner) -> None:
         [
             ArtifactRequirements(
                 [
-                    Coordinate(
-                        group="com.fasterxml.jackson.core",
-                        artifact="jackson-databind",
-                        version="2.12.1",
-                    )
-                    .as_requirement()
+                    ArtifactRequirement(coordinate=Coordinate(                         group="com.fasterxml.jackson.core",                         artifact="jackson-databind",                         version="2.12.1",                     ))
                     .with_extra_excludes("com.fasterxml.jackson.core:jackson-core")
                 ]
             ),
@@ -678,12 +678,7 @@ def test_missing_entry_for_transitive_dependency(rule_runner: RuleRunner) -> Non
         [
             ArtifactRequirements(
                 [
-                    Coordinate(
-                        group="org.apache.hive",
-                        artifact="hive-exec",
-                        version="1.1.0",
-                    )
-                    .as_requirement()
+                    ArtifactRequirement(coordinate=Coordinate(                         group="org.apache.hive",                         artifact="hive-exec",                         version="1.1.0",                     ))
                     .with_extra_excludes(
                         "org.apache.calcite:calcite-avatica",
                         "org.apache.calcite:calcite-core",
@@ -710,11 +705,7 @@ def test_missing_entry_for_transitive_dependency(rule_runner: RuleRunner) -> Non
 def test_failed_to_fetch_jar_given_packaging_pom(rule_runner: RuleRunner) -> None:
     reqs = ArtifactRequirements(
         [
-            Coordinate(
-                group="org.apache.curator",
-                artifact="apache-curator",
-                version="5.5.0",
-            ).as_requirement()
+            ArtifactRequirement(coordinate=Coordinate(                 group="org.apache.curator",                 artifact="apache-curator",                 version="5.5.0",             ))
         ]
     )
 
@@ -731,16 +722,22 @@ def test_force_version(rule_runner):
     # first check that force_version=False leads to a different version
     reqs = ArtifactRequirements(
         [
-            Coordinate(
-                group="org.apache.parquet",
-                artifact="parquet-common",
-                version="1.13.1",
-            ).as_requirement(),
+            ArtifactRequirement(coordinate=Coordinate(                 group="org.apache.parquet",                 artifact="parquet-common",                 version="1.13.1",             )),
             Coordinate(
                 group="org.slf4j",
                 artifact="slf4j-api",
                 version="1.7.19",
             ).as_requirement(),
+        ]
+    )
+    reqs = ArtifactRequirements(
+        [
+            Coordinate(
+                group="org.apache.parquet",
+                artifact="parquet-common",
+                version="1.13.1",
+            ).as_requirement(),
+            ArtifactRequirement(coordinate=Coordinate(                 group="org.slf4j",                 artifact="slf4j-api",                 version="1.7.19",             )),
         ]
     )
     entries = rule_runner.request(CoursierResolvedLockfile, [reqs]).entries
@@ -753,17 +750,26 @@ def test_force_version(rule_runner):
     # then check force_version=True pins the version
     reqs = ArtifactRequirements(
         [
-            Coordinate(
-                group="org.apache.parquet",
-                artifact="parquet-common",
-                version="1.13.1",
-            ).as_requirement(),
+            ArtifactRequirement(coordinate=Coordinate(                 group="org.apache.parquet",                 artifact="parquet-common",                 version="1.13.1",             )),
             dataclasses.replace(
                 Coordinate(
                     group="org.slf4j",
                     artifact="slf4j-api",
                     version="1.7.19",
                 ).as_requirement(),
+                force_version=True,
+            ),
+        ]
+    )
+    reqs = ArtifactRequirements(
+        [
+            Coordinate(
+                group="org.apache.parquet",
+                artifact="parquet-common",
+                version="1.13.1",
+            ).as_requirement(),
+            dataclasses.replace(
+                ArtifactRequirement(coordinate=Coordinate(                     group="org.slf4j",                     artifact="slf4j-api",                     version="1.7.19",                 )),
                 force_version=True,
             ),
         ]
